dot/requirements_data_confidence.txt
agate==1.6.1
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
asn1crypto==1.4.0
attrs==21.2.0
azure-common==1.1.27
azure-core==1.21.0
azure-storage-blob==12.9.0
Babel==2.9.1
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.12b0
bleach==4.1.0
boto3==1.20.20
botocore==1.23.20
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
cycler==0.11.0
dbt-core==1.0.0
dbt-extractor==0.4.0
dbt-postgres==1.0.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.28.3
google-api-core==1.31.2
google-auth==1.35.0
google-cloud-bigquery==2.31.0
google-cloud-core==1.7.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
great-expectations==0.12.10
grpcio==1.42.0
hologram==0.0.14
idna==2.10
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.0
jedi==0.18.1
Jinja2==2.11.3
jmespath==0.10.0
json-rpc==1.13.0
jsonpatch==1.32
jsonpointer==2.2
jsonschema==3.1.1
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==21.8.0
kiwisolver==1.3.2
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
matplotlib==3.5.0
matplotlib-inline==0.1.3
minimal-snowplow-tracker==0.0.2
mistune==0.8.4
msgpack==1.0.3
msrest==0.6.21
mypy-extensions==0.4.3
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.6
numpy==1.21.4
oauthlib==3.1.1
openpyxl==3.0.9
oscrypto==1.2.1
packaging==20.9
pandas==1.3.4
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.23
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.12.0
Pygments==2.10.0
PyJWT==2.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytimeparse==1.1.8
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
snowflake-connector-python==2.5.1
SQLAlchemy==1.3.24
sqlparse==0.4.2
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tomli==1.2.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
